requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
dpkt==1.9.8
filelock==3.9.0
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==15.0.7
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.3.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
velodyne-decoder==2.3.0
Werkzeug==2.3.3
